Moved data_converter code in H5Slice_traits_misc.hpp into its own file H5Converter_misc.hpp for reuse.
